[LABS-484] Extract clawback functionality into its own module
